blur sensitive headers in devtools
Some phishing attempts are getting targets into video calls to
"verify" themselves by showing info from devtools. They are assured
that they will not be pasting anything into devtools, and this is
true, the scammer just reads the authorization headers in the
screenshare. This CL blurs sensitive headers by default. This does
not affect the underlying text or the ability to copy-paste it.

Change-Id: I2ad7411c745236f73121f3f9158f911ca465c967