Handle response headers overrides defined in `.headers` files
This CL enables DevTools to override a request's response headers.
Header overrides are defined in local '.headers' files.

DevTools informs the Chromium backend about which requests to intercept
via CDP's `Fetch.enable`. The header overrides defined in '.headers'
files are taken into account when generating these interception
patterns.

When a request is being intercepted and DevTools is notified via
`Fetch.requestPaused`, DevTools checks if there are header overrides
which match this request. If so, it assembles the overridden response
headers by merging the response headers as they were received with the
local overrides defined by the user. The merged response headers are
then sent to the Chromium backend via `Fetch.fulfillRequest`.

Design Doc:
https://docs.google.com/document/d/1H45xqNnxx6B1oDbf3_4W-dJEAX-vDkDs0u8tE0tQ_H8/edit

Bug: 1288023
Change-Id: I07ada375b0ede7915dac8cbb39ee31ee63b3c328