[console] Always show the source location for SyntaxErrors
This CL changes how we render SyntaxErrors. For SyntaxErrors with
available ExceptionDetails, we add a single 'pseudo' stack frame that
can link to the source location of the error.

This way the user doesn't loose information when catching a
SyntaxError and printing it via `console.*` functions.

Screenshot: https://imgur.com/a/TWQ1nC7

R=bmeurer@chromium.org, pfaffe@chromium.org

Bug: 1066909
Change-Id: I73bd3b655f03fd10e2b5bcbe5e8c8090c53fa684