Ensure NetworkRequest is updated with overridden response headers
If there are both a `Network.responseReceived` and a
`Network.responseReceivedExtraInfo` event for a `NetworkRequest`,
DevTools previously always used the extraInfo event for populating
response header information.

When response headers are overridden with `Fetch.fulfillRequest`, this
does not generate an updated `Network.responseReceivedExtraInfo` event,
only a new `Network.responseReceived` is generated. This CL ensures
that DevTools keeps track of which requests have been overridden, so
that in these cases DevTools updates its internal representation of the
headers based on the updated `Network.responseReceived` event.

Bug: 1288023
Change-Id: I5c3c65f04c3cd17a042e2365b39623525bb8224f