Add --mocha-fgrep to unit tests
This CL adds the ability to run:

```
npm run auto-unittest -- --mocha-fgrep=breadcrumb
```

To the Karma unit test suite to mirror the similar flag available in the
interactions and e2e test runner script. This is also why it's named
`fgrep`, as that's the same flag as used in the other script, and we
should be consistent.

Bug: none
Change-Id: Ic119f7186e0e97c83e00bf92a66d69181c931a8b