[e2e] Load extensions from the same origin as the DevTools frontend
This CL is a preparatory refactoring. In the future, we want to load
DevTools from a different origin than `localhost` to increase
isolation between test runs (e.g. no shared window.localStorage).

To get there, we have to change how the extension tests load the
actual extension code. The extension must be served from the same
origin as the DevTools frontend. Currently this CL is a no-op, as
both resources and the DevTools frontend are served from 'localhost'.

R=pfaffe@chromium.org

Bug: 1297458
Change-Id: Iedee09d8cdd992f64132add78537428e9e4ab7ae