[e2e] Load DevTools on a different origin than the target page
This CL changes e2e test to load the DevTools frontend on the
'devtools-frontend.test' origin instead of using 'localhost'.
'localhost' is used by the target page. If we also load DevTools via
'localhost' than the two pages share window.localStorage and are
considered "same-origin", which hardly reflects reality.

To enable this CL, we have to change the hosted-mode server to set
some CORS headers. More specifically, we allow the DevTools frontend
to request arbitray resources from the hosted-mode server, while
target pages have to be explicit in their ".headers" or
".rawresponse" files.

R=jacktfranklin@chromium.org, pfaffe@chromium.org

Bug: 1297458
Change-Id: Ie18069e2effcc53cfd10a19296dc5d5c74b40e17