Skip RequestTimingView refresh if min timing boundary hasn't changed
The contents of the network request timing panel are recreated any time
the network panel start or end time boundaries change. If there is a
service worker respondWith dropdown open, then it appears to close
because the whole view (including the dropdown) is recreated. On sites
with lots of traffic where the boundary changes frequently, this makes
it difficult to keep the dropdown open.

Before: https://imgur.com/a/kXmgv8C

Since the timing panel only depends on the minimum time boundary, I've
updated the RequestTimingView to cache the previous value and only do a
refresh if the minimum boundary has changed since the last boundary
update.

I've verified that this fixes the bug and that the dropdown stays
open now.

Bug: 1298953
Change-Id: I2b4fa2e9b344d9beac5fd08df95c2b355216560f