[e2e] Use pre-allocated DevTools and target tabs
This CL changes the conductor to use new DevTools frontend and target
tabs for each individual test. A pool prepares a set of frontend/target
tab pairs in the background so the conductor doesn't need to waste time
resetting tabs in between tests.

For debugging e2e tests we disable the pool and use on-demand tab
creation instead. It would be annoying to deal with ~12 open tabs
and trying to find the right ones in a debugging session.

R=jacktfranklin@chromium.org

Bug: 1297458
Change-Id: I6b6c4a15e5d83f7e2bdc6082972993df179a5d87