[console] Correct autocompletion for arrays.
With the CodeMirror 6 migration, autocompletion regressed for array-like
objects, where we'd offer completions from the Array.prototype, but not
from the actual object (and its prototype chain) itself. The core
problem here is that for array(-like) objects, the front-end wouldn't
load the object properties itself, but instead loads the properties from
the Array.prototype (or the Uint8Array.prototype), even if its not
present in the objects prototype chain. In other words, we pretend that
the object has the properties of the Array.prototype, even when it
doesn't and at the same time don't show its actual properties for
completion. This doesn't make any sense, so let's get rid of the special
casing all together and instead ask the autocompletion to not return any
indexed properties (which was the whole reason for doing this in the
first place).

Drive-by-fix: Also support autocompletion for primitive symbols.

Fixed: chromium:1297101
Change-Id: I19c6abe1d23d5c7765e5ef1165e48b31aa3cf3af