Add back triangles for code folding
With CodeMirror 6 we were using the default CodeMirror styled arrows,
whereas with CodeMirror 5 we had a custom version. After user feedback,
we opted for adding back the original arrows and move the code folding
buttons to the right of the line numbers.

We use the expand/collapsed arrows in various places, but they use the
legacy UI `smallIcons.svg` built with webkit mask images. To avoid
relying on the legacy icons, I opted for using the modern `Icon.ts` by
introducing new arrows for collapsing and expanded. These arrows should
be visually equivalent to the legacy versions. Please see the screenshot
below for the new CodeMirror 6 version.

Image: https://imgur.com/a/jyHQFkc

R=petermueller@chromium.org,yangguo@chromium.org

Fixed: 1299364
Change-Id: Ic8921229c3657fa4ad404d7c38a80ccc192e1319