Fix Media>Messages dropdown not highlighting properly
The custom color styling here overrides the highlighting CSS that's
coming from the SoftDropdown styles, and messes up the highlighting as
the user uses their keyboard to navigate the list.

We don't need these styles; they are applied in SoftDropdown which also
accounts for the highlighting via keyboard nav etc.

Fixed: 1302956
Change-Id: I74a78b8d790ea93148c2fd3c51c684612f3a2b2c