[debugger] Fix stepping at the end of a sourcemapped script.
Sourcemap entries don't contain end positions, but assume that they just
span until the next entry, which works fine, except for the last entry
in a sourcemap for which we don't know the end line/column. Previously
we put the pair Infinity:0 there, but the Infinity cannot be transfered
as `line` in a `ScriptPosition`, since the CDP has type int32 for this
field.

This changes the sentinel position to use the maximum signed 32-bit
integer for both end line and end column for the last sourcemap entry,
which is definitely outside any valid script and can safely be
transfered to V8.

Fixed: chromium:1305956
Change-Id: I7481fce7579464169397f601b13c57ee817ea105