[sources] Always show correct files in the text editor.
This undoes the part of https://crrev.com/c/2107522 which reveals the
canonical UISourceCode instead of the actual UISourceCode in the Sources
panel.

This behavior was very confusing and made it impossible to reveal the
correct files in case of an URL clash (which unfortunately is quite easy
to run into with modern web tooling, and might even be outside of the
control of the web developer). This is illustrated with a trivial E2E
tests that has two different index.js files, one for the main frame and
one for a worker, and where before DevTools front-end would always show
the index.js from the main frame (the "canonical one") even when the
developer explicitly clicks on the index.js in the worker.

Fixed: chromium:111297
Bug: chromium:670180
Change-Id: Ie078bf79e47d1fa9a09341d39ffe1abb88909f81