Do not register the same layout invalidation twice
This CL prevents an error to be thrown when the performance panel
tries to match `StyleInvalidatorInvalidationTracking` with
`ScheduleStyleInvalidationTracking` trace events. Matching these two
allows to provide more details about the initiators of style
recalculations in the renderering pipeline in a trace, and is part of an
unstable feature implemented eight years ago. Since the its
implementation (in https://codereview.chromium.org/731293006),
the feature is run under a flag for unstable experiments, so the events
in question haven't been fully supported by DevTools to date.

The fix consists in not adding `ScheduleStyleInvalidationTracking`
events as layout invalidations twice. Currently, the events are added
once they are handled from the trace and another time when a
`ScheduleStyleInvalidationTracking` is handled. When the latter happens,
the style recalculation to which the initiators want to be added is set
to null, as if another frame has started, which eventually causes the
error to be thrown. As such this CL removes the second registration to
prevent the throw.

Finally, the error logged when no match is found for a
`StyleInvalidatorInvalidationTracking` event is omitted as this is a
common case, according to tested traces, and it causes too much noise in
the console.

Fixed: 1306445
Change-Id: I6eb5fb6dc3855bc264b3b91cc47d7a8a37290376