Copy requests as Python in the Network tab
Allows users to copy a request as a Python function using the
popular https://requests.readthedocs.io/ library from the
Network tab.

If everyone started implementing this for their favorite language
that menu would get quite long. On the other hand,
https://curlconverter.com supports a dozen languages so developers
cleary want this. We could add another sub-menu.

I use Python for scraping and I usually start by copy and pasting
a request as cURL and using curlconverter to convert the cURL
command to Python. This is a pretty popular website, I think a lot
of people are doing this and I wanted to remove the extra Ctrl-tab,
Cmd-V, Tab, Cmd-C by doing it in Chrome directly.
It should be more reliable to go
request -> Python
directly instead of
request -> cURL -> *JavaScript bash parser* -> Python

Change-Id: Id78564ea4353f13703ae2f1cebf305fbea8cf448