Remove rejector warning from RenderCoordinator
`this.#rejectors` is a WeakMap of callback funcs => rejector functions
from promises; if it cannot find a value it's probably because the
reference has been dropped as per how a WeakMap behaves, so let's just
silently do nothing and move on.

This prevents some slight noise in the unit tests.

Bug: none
Change-Id: I703da4c8c5ba10053e2bc2602e5539566bc32b7b