Handle falsy origin when accessing storage for Lighthouse
`mainTarget.inspectedURL()` can sometimes return an empty string. This
can cause `Storage.getUsageAndQuota` to return an error. Our code throws
an error because this error does not contain the requisite properties
expected on the normal result.

This CL handles issue more gracefully by exiting early if the origin
is falsy.

Bug: null
Change-Id: I71f5c105dcde4f80983f0e37eb440e51392cde2e