Only style selected TreeNodes when the outline is focused
This CL fixes an issue where we render a tree-outline and immediately
see the first node selected and styled as such. This is misleading
because it looks to the user like it's focused, but it's not (unless we
have auto-focused that element).

To fix this we'll apply the selected styles only when the <ul> has
:focus-within.

Bug: none
Change-Id: I1b93abe4607d295d79d76db06178f219d69e2d10