Test ScriptFormatter.format for workspace sync
How to test:
- Open DevTools on DevTools (DevTools1 = targeting page, DevTools2 = targeting DevTools 1)
- In DevTools2, place a breakpoint InplaceFormatterEditorAction.contentLoaded on the line with "await Formatter.ScriptFormatter.testFormat();"
- In DevTools1, Go to sources tool and press the pretty print button on a source file to trigger breakpoint
- Jump into Formatter.ScriptFormatter.testFormat(), look at result, etc.

The ScriptFormatter.testFormat() method has hard coded strings from what we would expect from a stylesheet.
This is the base testing stylesheet: https://imgur.com/9MZB8sA

The three test cases are incrementing a property, removing a property and adding a property. The return format for the funciton is:
[
	[originalString, formattedString, expectedFormattedString],
	...,
	...,
]

This is the result: https://imgur.com/mU2d7PM
(right click the image and click "open image in new tab" to zoom into the screenshot)

Change-Id: I0c30a948402aa91caf3fa710ab8ab9d117819188