[code-health] Add formatter helpers to parse the devices list
A Python script used to parse the emulated devices list to generate a
C++ header and source files to keep this list up-to-date for
chromedriver. However, this list has been moved many times and is no
longer easily parseable by that Python script.

This CL introduces formatter helpers to make it easier to parse that
list and generate the C++ code at build time rather than using a
manual process.

See https://crrev.com/c/3558793 for the Python script.

Bug: chromedriver:3772
Change-Id: Ifdb23fffa54f1d644c0365329d3d2aecae571152