Use UISourceCode for html script first for inline scripts  w/o sourceURL
The breakpoints for inline scripts without source urls need to be
set on the UISourceCode of the html file. However, this may be
appearing late in time, and if it is not yet available we cannot
block on it (since we would deadlock), but need to use the v8 script
of the inline scirpt.

However, if the UISourceCode of the html file is already available,
use it.

Bug: chromium:1229541, chromium:1133307,
Change-Id: I698421767f0da213547fcc223a748904c2b46488