Fix host bindings
The CL https://crrev.com/c/3565361 added a method to the host bindings
stub to reset the cached stub data. This does not really work as we
explicilty check on the C++ side that the host bindings are fully
implemented.

This CL is a forward fix that moves the resetting out of the
bindings stub. The `#urlsBeingSaved` Map is not cleared, given
that its in-accessible by the outside anyway.

Roll failure: https://crrev.com/c/3564055.

R=dsv@chromium.org

Bug: None
Change-Id: I94fb41ded43540b742883b96b358f3d3b727dfac