[sources] Disable autocompletion for the breakpoint editor.
Currently we don't have a good solution to provide sensible completions
for the breakpoint editor, so either the developer ends up with only
completions from the global object (debugger isn't paused), or the
developer ends up with completions from the currently paused scope,
while trying to set a conditional breakpoint or logpoint in a completely
different function. Ideally we'd perform some kind of syntactic analysis
here, and also pay special attention (and warn developers about
potential issues) in the case of sourcemapped code.

For now, the autocompletion is primarily an annoyance, and so it's
better to disable it for now in this case, until we have a viable
solution that doesn't get in the way.

Fixed: chromium:1312374
Bug: chromium:1276960
Change-Id: I863ed80dd7cd501c5faa9e88acd1453f0639aece