Remove resolver before calling handler in render coordinator
Sometimes we're receiving "Resolver not found" errors for some
specific execution of the consecutive renders of same component

It's challenging to replicate the exact
scenario in unit tests so we're not adding
a test here.

However, this change fixes the issue. Here are the execution paths
for problematic renders for before
and after cases:

Before
* Scheduled render for 1st render
* Add pending renders for 1st render
* Set resolver for 1st render
* Wrapped callback called for 1st render
* Scheduled render for 2nd render
* Add pending renders for 2nd render
* Set resolver for 2nd render
* Resolved for 1st render
* Removed resolver for 1st render
* Write completed for 1st render
* Wrapped callback called for 2nd render
-> Unable to locate resolver for 2nd render

After
* Scheduled render for 1st render
* Add pending renders for 1st render
* Set resolver for 1st render
* Removed resolver for 1st render
* Wrapped callback called for 1st render
* Scheduled render for 2nd render
* Add pending renders for 2nd render
* Set resolver for 2nd render
* Resolved for 1st render
* Write completed for 1st render
* Removed resolver for 2nd render
* Wrapped callback called for 2nd render
* Resolved for 2nd render
* Write completed for 2nd render

Essentially, what happens in this execution is:
* 1st render happens (wrapped callback is called)
* Before it is resolved and its resolver is removed
2nd render is scheduled which adds
the resolver to the map with the same key
* Resolver is called and removed for 1st render
* Resolver is tried to be retrieved for 2nd render and it fails

In here, we separate
removing the resolver and calling it
into different times of execution.

Though I wonder if there can be a way
to replicate this in tests.

Change-Id: Ic65feb1adb54792e1f9c6d10efb5f6608875170d