Add repro for stuck Service Worker pending request
The main request for a Service Worker scripts currently gets stuck in
the "(pending)" state.

This test will currently fail like:

```
1) network/network-datagrid_test.ts: The Network Tab
      shows the main service worker request as complete:

    AssertionError: expected { Object (status, time, ...) } to have property 'status' of '200OK', but got '(pending)'
    + expected - actual

    -(pending)
    +200OK
```

The issue appears to be rooted in the misalignment of network request ids
between the inial network CDP event for the request compared with
subsequent events.

Running this test with an executable built from https://chromium-review.googlesource.com/c/chromium/src/+/3544685
passes.

Bug: 1304795
Change-Id: I14f8d64d58fad16e4ddb0acf2058ed4c613c5500