Color picker switches to the new EyeDropper API
The EyeDropper API[1] has been enabled by default in Chrome 95[2]. Using the new API,we can draw colors outside of the Chrome window.

Added the `EYEDROPPER_COLOR_PICKER` experimental flag, which is on by default,
and the user can turn off this flag to switch to the old version of the color picker.

[1]: https://github.com/WICG/eyedropper-api/
[2]: https://bugs.chromium.org/p/chromium/issues/detail?id=897309#c35

Bug: 1245191
Change-Id: I05f8628322ab5d314cca640fb269b865821f2ae4