[e2e] Fix and reenable Multi-Workers test cases.
This refactors the test cases to avoid racing with the Revealer in the
devtools-frontend by explicitly opening the desired files ahead of time
and only checking that DevTools doesn't reveal a second copy of the
file.

Drive-by-fix: Also fix the checking for hitting the correct breakpoint
by aligning the testing logic here with how it's done in other places
(the old CodeMirror execution line detection no longer works anyways).

Bug: 670180
Change-Id: I382a7956bef66df53d3fb95571cb31cc086d9105
Fixed: 1157455