Use single table for legacy DataGrid component
This change converts the separate <table> that was used for the
DataGrid's header row into a <thead> section in the data table. This
results in a single <table> element which is easier for screen reader
users to navigate.

The header has CSS position:sticky so that it remains in view when
scrolling through long data sets. To get the cleaner DOM structure,
and ensure scrolling continued to work, I had to set the whole table
to overflow-y: overlay. This means that the scrollbar now extends up
into the header row which is a visual change from before. See before
and after screenshots below.

I looked into setting the overflow property on the
<tbody> element instead, but this would only make the <tbody>
scrollable if the <tbody>'s display was set to 'block'. Doing that
keeps the scrollbar out of the header area as before, but it breaks
the overall table layout and prevents the cells from using the
widths defined in <colgroup>, which means the resizer UI doesn't work.
This could potentially be fixed with more work in the typescript, but
its not clear this would be worth the effort. The change I'm sharing
here trades the small visual regression for an accessibility win with
as little code churn as possible.

Before:
https://imgur.com/a/9ffDg3J

After:
https://imgur.com/a/iPcnaw7

Bug: 1316147
Change-Id: I30638eb512cdf91e9e5b6383bee8eb1289527311