Lint for checkbox stylesheet adoption in Lit components
This CL introduces (but does not enable) an ESLint rule to check that if
a Lit component renders a checkbox, that it also imports and adopts the
common `Input.checkboxStyles` into the component's shadow root.

We can also roll this rule out to the text inputs too, but I wanted to
start with just one and see how it goes. I will follow up this CL with a
CL to enable this rule.

Bug: none
Change-Id: Ibed7109d5948c8be3f4c9d376738290137e444ff