Fix local override filename handling for special characters
This CL fixes local overrides for URLs which contain special characters.
When a resource is saved for overrides, a local path and filename are
generated from the resource's URL. Because not all characters are
allowed to be used in filenames, these characters need to be escaped
(via URL-encoding). The most common cases are '?', which is used for URL
query params and ':', which is used for the port number. Both '?' and
':' are not allowed within file names on Windows OSes.

Previously, DevTools was hiding this complexity of encoded file names
from the user by decoding the file names for user-facing strings. This
could lead to unexpected behaviour, e.g. when renaming files for which
the actual filename does not match the user-facing filename. With this
CL the actual (encoded) filename is shown to the user.

Bug: 1308257
Change-Id: I0303af33502e4e5768e0ccdb44d495b6a2b62190
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3579160
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
(cherry picked from commit 81477453b5ffc2dfdac05c5230fa5422ae86b1e8)