Switch builders running chromium_integration to chromium_trybot.
chromium_trybot is nearly identical to chromium_integration. The two
recipes call different *_steps functions, but the way they are called by
the two recipes results in equivalent behavior. The only other
difference between the two recipes is that chromium_trybot has an early
assertion that the build has a CL to test, which is a quality-of-life
improvement that results in a clearer error if someone attempts an led
run and forgets to set a CL.

Change-Id: Ia0922e81812d60145b9488030454aa9009cbaf0a