Restore ordering of pseudos in style pane
When trying to re-enable the test disabled in [1], I noticed that [2]
inadvertently changed the order in which highlight pseudos are listed
relative to other pseudos in the style pane.

This change restores the initial ordering, which is that ::before
pseudos are listed first followed by all other pseudos in
alphabetical order.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3584106
[2] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3603063

Bug: 1024156, 1164461
Change-Id: Ic2758995593da20739b58245a7d48243963fd18a